refactor(ui): DOMA-6670 refactored Dropdown and DropdownButton
