+++ packages/ui/src/components/Button/style.less
@@ -323,35 +323,62 @@ a.condo-btn {
   }
 }
 
-.condo-dropdown-trigger {
+.condo-dropdown-trigger.condo-btn-extended {
+  padding: 0;
   cursor: default;
-
-  &.condo-btn > .condo-btn-label {
-    margin-right: auto;
-    margin-left: auto;
-  }
-
-  .condo-btn-vertical-divider {
-    z-index: 1;
-    margin-right: @condo-global-spacing-16;
-    margin-left: @condo-global-spacing-20;
-    opacity: 0.1;
-  }
-
-  &.condo-btn-secondary.condo-btn-dangerous {
-    .condo-btn-vertical-divider {
-      color: @condo-global-color-red-5;
-    }
-
-    &:hover {
-      .condo-btn-vertical-divider {
-        color: @condo-global-color-red-7;
-      }
-    }
-  }
 
   .condo-btn-ellipsis {
     z-index: 1;
     display: inline-flex;
+    padding: 12px 20px 12px 0;
+  }
+
+  .condo-btn-label {
+    max-width: calc(100% - 20px - 20px);
+    margin-right: auto;
+    margin-left: auto;
+    padding: 12px 0 12px 20px;
+    overflow: hidden;
+    text-overflow: ellipsis;
+  }
+
+  .condo-btn-text {
+    display: inline-flex;
+    align-items: center;
+    width: 100%;
+    height: 100%;
+  }
+
+  .condo-btn-vertical-divider {
+    width: 1px;
+    height: 100%;
+    margin-right: @condo-global-spacing-16;
+    margin-left: @condo-global-spacing-20;
+    background-color: @condo-global-color-white;
+    opacity: 0.1;
+  }
+
+  &.condo-btn-secondary > .condo-btn-text {
+    .condo-btn-vertical-divider {
+      background-color: @condo-global-color-black;
+    }
+  }
+
+  &.condo-btn-secondary.condo-btn-dangerous > .condo-btn-text {
+    .condo-btn-vertical-divider {
+      background-color: @condo-global-color-red-5;
+    }
+  }
+
+  &.condo-btn-secondary:not(.condo-btn-dangerous):hover > .condo-btn-text {
+    .condo-btn-vertical-divider {
+      background: @condo-global-color-brand-gradient-5 border-box;
+    }
+  }
+
+  &:hover.condo-btn-secondary.condo-btn-dangerous > .condo-btn-text {
+    .condo-btn-vertical-divider {
+      background-color: @condo-global-color-red-7;
+    }
   }
 }
