+++ packages/ui/src/components/Dropdown/dropdown.tsx
@@ -1,6 +1,7 @@
 import { Dropdown as DefaultDropdown, DropdownProps as DefaultDropdownProps } from 'antd'
-import { isFunction } from 'lodash'
 import get from 'lodash/get'
+import isFunction from 'lodash/isFunction'
+import { MenuInfo } from 'rc-menu/lib/interface'
 import React, { useCallback } from 'react'
 
 import { extractChildrenContent, sendAnalyticsClickEvent } from '../_utils/analytics'
@@ -11,7 +12,7 @@ export const DROPDOWN_CLASS_PREFIX = 'condo-dropdown'
 export type DropdownProps = Omit<DefaultDropdownProps, 'visible' | 'prefixCls' | 'overlayStyle'>
 
 const Dropdown: React.FC<DropdownProps> = ({ menu, children,  ...props }) => {
-    const handleClickMenu = useCallback(async (info) => {
+    const handleClickMenu = useCallback(async (info: MenuInfo) => {
         const triggerId = get(children, 'props.id')
         const triggerStringContent = extractChildrenContent(children)
         const onClickMenu = get(menu, 'onClick')
